fix: resolve bundling issue with config imports
- Move config schema from root archivist.config.ts to src/config/schema.ts
- Update all imports to use internal path instead of external
- Keep archivist.config.ts as re-export for backward compatibility
- Rebuild CLI with fixed imports

This fixes the issue where archivist would hang when installed as a dependency
because the bundler couldn't properly resolve imports from outside the src directory.
